add rocksdb build tag in order not to build rocksdb default
