+++ storage/database/dynamodb.go
@@ -444,6 +444,10 @@ func (dynamo *dynamoDB) Meter(prefix string) {
 	dynamoBatchWriteTimeMeter = metrics.NewRegisteredMeter(prefix+"batchwrite/time", nil)
 }
 
+func (dynamo *dynamoDB) GetProperty(name string) string {
+	return ""
+}
+
 func (dynamo *dynamoDB) NewIterator(prefix []byte, start []byte) Iterator {
 	// TODO-Klaytn: implement this later.
 	return nil
