+++ storage/database/rocksdb_database_nobuild.go
@@ -1,0 +1,28 @@
+// Copyright 2023 The klaytn Authors
+// This file is part of the klaytn library.
+//
+// The klaytn library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The klaytn library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the klaytn library. If not, see <http://www.gnu.org/licenses/>.
+
+//go:build !rocksdb
+// +build !rocksdb
+
+package database
+
+import "errors"
+
+var ErrRocksDBNotBuilt = errors.New("rocksdb is not built")
+
+func NewRocksDB(path string, config *RocksDBConfig) (Database, error) {
+	return nil, ErrRocksDBNotBuilt
+}
