+++ storage/database/sharded_database.go
@@ -428,6 +428,15 @@ func (db *shardedDB) Meter(prefix string) {
 	}
 }
 
+func (db *shardedDB) GetProperty(name string) string {
+	var buf bytes.Buffer
+	for index, shard := range db.shards {
+		buf.WriteString(fmt.Sprintf("shard %d:\n", index))
+		buf.WriteString(shard.GetProperty(name))
+	}
+	return buf.String()
+}
+
 type shardedDBBatch struct {
 	batches    []Batch
 	numBatches uint
